gittools: Produce proper PowerShell-stylle output when required
Using WriteHost is usually not the way to output in PowerShell (read
internet to know why) but was chosen here because it gives visually
clear output for everyday use. However for programmatic use of
Invoke-Mr we want something easy to consume so add a -Table switch
which will yield a proper table as output.